Update Docker version and add explanation.
I noticed the version of docker installed on the stack's EC2 instance was inconsistent with the expected value on the assignment.

As I was going through the assignment, I noticed an opportunity to explain both why th e`-P` flag was used to launch the nginx container and the port mapping output in the `docker ps` command.